This edit cut the spec down and further developed the format we have chosen.
This is the last major single edit before we run through this as a group.   This edit  also split the the Use Cases from this core specification document. 

The edits have included:
- pulled in and refined further the core extensions from v.05
- Added core field definition/description section (TBD)
- Added core field format section (TBD)
- Cut Use case to new document
- Added Compliance by Default Scope
- Added an MVCR appendix for usability metrics

This edit did not include core definitions.  I have added notes for discussion points and rationale, the aim is to try and contain these discussions to the list as much as possible. 

Proposed way to proceed forward:
1.  Run through MVCR framework in calls
- note all comments, issues and note definitions that need work (as offline action items)
2. Address Framework Questions & 
3. Polish definitions on second run through
4. Complete Reference (sub- Documents)
5. Complete Demonstrator Use Case: Website based on MVCR (led by Iain)